+++ app/src/main/java/com/futsch1/medtimer/overview/EditEventSideSheetDialog.kt
@@ -1,4 +1,4 @@
-package com.futsch1.medtimer.new_overview
+package com.futsch1.medtimer.overview
 
 import android.annotation.SuppressLint
 import android.view.LayoutInflater
