+++ app/src/main/java/com/futsch1/medtimer/overview/VariableAmount.kt
@@ -1,4 +1,4 @@
-package com.futsch1.medtimer.new_overview
+package com.futsch1.medtimer.overview
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
