+++ app/src/main/res/layout/fragment_overview.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".new_overview.NewOverviewFragment">
+    tools:context=".overview.OverviewFragment">
 
     <com.kizitonwose.calendar.view.WeekCalendarView
         android:id="@+id/overviewWeek"
