+++ app/src/main/res/menu/bottom_nav.xml
@@ -2,8 +2,8 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/newOverviewFragment"
-        android:name="com.futsch1.medtimer.new_overview.NewOverviewFragment"
+        android:id="@+id/overviewFragment"
+        android:name="com.futsch1.medtimer.overview.OverviewFragment"
         android:contentDescription="@string/overview_tab_description"
         android:icon="@drawable/calendar2_event"
         android:title="@string/tab_overview" />
