+++ app/src/main/res/navigation/navigation.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/navigation"
     android:label="@string/tabular_view"
-    app:startDestination="@id/newOverviewFragment">
+    app:startDestination="@id/overviewFragment">
 
     <fragment
         android:id="@+id/medicinesFragment"
@@ -120,7 +120,7 @@
         android:name="com.futsch1.medtimer.preferences.DisplaySettingsFragment"
         android:label="@string/display_settings" />
     <fragment
-        android:id="@+id/newOverviewFragment"
-        android:name="com.futsch1.medtimer.new_overview.NewOverviewFragment"
+        android:id="@+id/overviewFragment"
+        android:name="com.futsch1.medtimer.overview.OverviewFragment"
         android:label="@string/tab_overview" />
 </navigation>
